+++ types/factories/IBondDepository__factory.ts
@@ -92,9 +92,9 @@ const _abi = [
     name: "debtDecay",
     outputs: [
       {
-        internalType: "uint64",
+        internalType: "uint256",
         name: "",
-        type: "uint64",
+        type: "uint256",
       },
     ],
     stateMutability: "view",
